requirements.txt
streamlit
torch
torchvision
opencv-python
Pillow
numpy
pandas
altair